api/requirements.txt

python-dotenv
fastapi[standard]
google-genai
pydantic